% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment-b2-mbeletsky.R
\name{group_then_sumz}
\alias{group_then_sumz}
\title{Group your data by a variable, then summarize unique observations in another variable}
\usage{
group_then_sumz(data, group_var, count_var)
}
\arguments{
\item{data}{A dataframe you are working with. I named this argument "data" to keep things simple and to the point, and for existing functions, it's common for the argument "data" to be a dataframe.}

\item{group_var}{The variable (column) you wish to group your data by. I named this argument "group_var" because it's the variable (var) you are wanting to group by (group) and to distinguish it easily from the other variable we are using as an argument.}

\item{count_var}{The variable (column) you wish to count unique observations for. I named this argument "count_var" because it's the variable (var) you are wanting to count unique observations in (count) and to distinguish it easily from the other variable we are using as an argument.}
}
\value{
Tibble that shows counts of unique observations within a variable, for each group you specified based on the grouping variable.
}
\description{
Group your dataset by one variable, then summarize counts of unique observations for another variable. Rows will be sorted from highest counts to lowest.
}
\examples{
library(datateachr)
group_then_sumz(vancouver_trees, neighbourhood_name, species_name)
library(palmerpenguins)
group_then_sumz(penguins, island, species)
}
